+++ src/aim/config/__init__.py
@@ -1,2 +1,2 @@
-from aim.config.cfg_codecommit import CodeCommitConfig
+