+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Changelog for aim
 - Moved yaml.py to aim.core
 - Refactored S3 Controller
 - Ported Route53 config to the model
-
+- Ported CodeCommit config to the model
 
 0.6.0.dev0 (2019-06-21)
 -----------------------